move add_thread code to lib/rex/post/meterpreter/extensions/peinjector/peinjector.rb
